+++ core/src/main/java/com/linecorp/armeria/common/CommonPools.java
@@ -19,7 +19,6 @@ package com.linecorp.armeria.common;
 import java.util.concurrent.ScheduledExecutorService;
 
 import com.linecorp.armeria.client.ClientFactoryBuilder;
-import com.linecorp.armeria.common.util.BlockingTaskExecutor;
 import com.linecorp.armeria.common.util.EventLoopGroups;
 import com.linecorp.armeria.server.ServerBuilder;
