+++ core/src/main/java/com/linecorp/armeria/server/AnnotatedServiceBindingBuilder.java
@@ -34,7 +34,7 @@ import com.linecorp.armeria.common.DependencyInjector;
 import com.linecorp.armeria.common.SuccessFunction;
 import com.linecorp.armeria.common.annotation.Nullable;
 import com.linecorp.armeria.common.annotation.UnstableApi;
-import com.linecorp.armeria.common.util.BlockingTaskExecutor;
+import com.linecorp.armeria.common.BlockingTaskExecutor;
 import com.linecorp.armeria.internal.server.annotation.AnnotatedServiceElement;
 import com.linecorp.armeria.internal.server.annotation.AnnotatedServiceExtensions;
 import com.linecorp.armeria.internal.server.annotation.AnnotatedServiceFactory;
